backend/requirements.txt
Django==5.1.4
djangorestframework==3.14.0
django-cors-headers==4.3.1
boto3==1.34.14
django-storages==1.14.2
python-dotenv==1.0.0
gunicorn==21.2.0
mysqlclient==2.2.6